Modified LocalStore to use the new database schema for messages
A lot more changes are required to meet the goal of being able to
reconstruct byte-identical copies of messages. But the current code is
too difficult to modify to reach that goal. And since I want to
replace LocalStore with a ContentProvider-based solution anyway work
on this branch is abandoned. I should be able to reuse most of the
code, though.

I am sorry for the large commit, but I lack the energy to decompose it
into smaller parts :(