Delete local messages after copy/move if we didn't get the new UIDs
We first copy/move messages locally and then issue the command to the
server. If we can't get the new UIDs for the messages in the target
folder on the server, we have to delete the local messages and then
(re)download them the next time the folder is synchronized.